requirements.txt
asgiref==3.4.1
authentication==1.1.0
autocommand==2.2.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-filter==2.4.0
django-heroku==0.0.0
django-nine==0.2.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
djoser==2.1.0
dnspython==2.1.0
elastic2-doc-manager==1.0.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.3
importlib-resources==5.2.2
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mongo-connector==3.1.1
mongoengine==0.23.1
oauthlib==3.1.1
passlib==1.7.4
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pymongo==3.12.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
WTForms==2.3.3
zipp==3.5.0
